+++ tokio/src/net/mod.rs
@@ -49,8 +49,6 @@ cfg_net_unix! {
     pub use unix::datagram::socket::UnixDatagram;
     pub use unix::listener::UnixListener;
     pub use unix::stream::UnixStream;
-
-    pub mod pipe;
 }
 
 cfg_net_windows! {
