+++ tokio/tests/io_unix_pipe.rs
@@ -1,8 +1,8 @@
 #![cfg(feature = "full")]
 #![cfg(unix)]
 
+use tokio::io::unix::pipe;
 use tokio::io::{AsyncReadExt, AsyncWriteExt, Interest};
-use tokio::net::pipe;
 use tokio_test::task;
 use tokio_test::{assert_err, assert_ok, assert_pending, assert_ready_ok};
